chore: simpler ber2der with length check
Do not build an intermediate tree presentation of BER format in ber2der. Rather
write the data directly to the output array patching the length header as
necessary.

Protect against int32 overflow on 32-bit CPU when calculating the BER content
length.

Detect and report invalid BER when a fixed length of the sequence does not match
the actual length of children.